feat: refactor test namespaces and add missing test annotations in AuthControllerTest and UserControllerTest
